Team synchronization bug fixed

+++ planIt/src/main/java/planit/project/services/ChatService.java
@@ -58,11 +58,11 @@ public class ChatService {
 		return null;
 	}
 
-	public List<MessageDTO> getMessages(Integer teamId) {
+	public List<MessageDTO> getMessages(Long teamId) {
 
 		List<MessageDTO> messageDTOs = new ArrayList<MessageDTO>();
 
-		Optional<Team> team = teamRepository.findById(teamId.longValue());
+		Optional<Team> team = teamRepository.findById(teamId);
 		if (team.isPresent()) {
 			List<UserMessage> messages = userMessageRepository.findAllByTeam(team.get());
 			if (messages != null) {
